+++ CHANGELOG.md
@@ -28,3 +28,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Remove IDAT configurable
 - Gender converter and birthday converter
 - Add Pseudonym to data mutations
+- Pass fhir results of python script to blaze exporter